requirements.txt
Flask==2.3.0
Flask-Mail==0.9.1
gunicorn==20.1.0
Pillow>=11.0.0
python-dotenv==1.0.0
xhtml2pdf==0.2.8

# Notes:
# - `Flask` and `Flask-Mail` are required by `app.py` and Netlify function.
# - `Pillow` is required by `create_favicon.py` (uses `PIL`).
# - `python-dotenv` is used to load `.env` files for local development (MAIL_USERNAME, MAIL_PASSWORD).
